Support reading and writing custom attributes on InterfaceImpl
WinRT uses custom attributes in winmd files that are declared for InterfaceImpl rows.
This is a breaking API change that we could not represent before.
There's other places that we do not support reading and writing custom attributes on,
like generic parameter constraints. Let's fix it when we need it.