[Performance] Improve AddMapping performance by using a List instead of an Array.
Mono.Cecil is used by mono-addins to scan all addin assemblies. In this
case, we request the types from a given assembly, which goes through the
nested types cache mapping.

A hot path in this case was AssemblyReader.AddMapping, which was doing a
lot of array resizing in the end.

Using mono-addins inside MonoDevelop would allocate 31MB of arrays.

MetadataReader.InitializeTypeDefinitions would allocate 57MB before, and
now it's down to 27MB.

The CPU time is also down now, from 117ms to 71ms.