Enable restoring of Cecil as a P2P reference in .NET Core
A NuGet bug (https://github.com/NuGet/Home/issues/4873) causes the
configuration not to be set during restore when using Cecil as a
ProjectReference of another project using the .NET Core cli tooling.

Until that bug is fixed, this workaround will correctly import
NetStandard.props by checking for the existence of an msbuild property
that is set during restore using the new tooling.