Add support for delay signing an assembly with only a public key file
Problem:
Mono.Cecil attempts to strong sign an assembly with a StrongNameKeyPair that only contains a public key. This throws an exception as a private key is required to complete the sign. However, it should be supported to place the public key into the AssemblyName without signing the assembly.

Fix:
Allow clients to pass in the public key.  This should be passed in alongside a StrongNameKeyPair.  If the strong name fails to calculate a PublicKey and throws the ArgumentException, set the PublicKeyBytes to the assembly instead. Furthermore, do not attempt to sign/strong name the assembly in this scenario.