Port Mono.Cecil to .NET Standard 2.0.
Changed the target runtime to .NET Standard 2.0.
Removed some conditional compilation ifs that were
removing the ability to strong name code.
Bring back code in BaseAssemblyResolver that builds on
.NET Standard 2.0, enabling the search of assemblies in the
GAC, allowing the code to find assemblies in the GAC, even when
referenced from a net461 DLL in .NET Framework full.