Improves thread safety of lazy initializations
The Interlocked.Exchange calls will ensure that the same object is
returned from each call even under a race.  That's not necessarily
required, but it should also make sure that no write re-ordering issues
occur on platforms that allow it.  And it also makes the possibility
of thread safety issues explicit.