Support custom streams with file name info
Many operations involved in rewriting debug symbols assume the rewritten
assembly base stream is a FileStream to get file name information. This
may not always be the case.

To cover this, this changes the way file name is resolved so that a
custom stream implementing `IHaveAFileName` can be understood by Cecil.

Signed-off-by: Simon Ferquel <simon.ferquel@unity3d.com>