Emit sequence points in ascending offset order
This is required by the Portable PDB specification, and the sequence
point collection in `MethodDebugInfo` does not guarantee it.

This also tweaks the writing of hidden sequence points to use two
writes of 0 as compressed unsigned integer, instead of a single
write of 0 as a plain short, to more closely match the specification.

Fixes #840.