Add the "allow byref-like" anti-constraint
This adds a new `AllowsByRefLike` flag to `GenericParameterAttributes`
which, unlike the others, is an anti-constraint (i.e. it removes
restrictions rather than adding them) added in C# 13.
Because of this, `AllowsByRefLike` was chosen as the corresponding bool
property on `GenericParameter` (instead of
`HasAllowByRefLineConstraint`; the constant also does not include
`Constraint` in its name).

At the same time, it adds the `NoSpecialConstraint` flag and a
corresponding `HasNoSpecialConstraint` property.